Rename serialize -> serializer
