+++ src/main.rs
@@ -7,7 +7,7 @@ mod engine;
 mod play;
 mod record;
 mod remote;
-mod serialize;
+mod serializer;
 mod setup;
 mod sparse_mat;
 mod train;
@@ -18,7 +18,7 @@ use crate::engine::board::*;
 use crate::engine::search::*;
 use crate::play::*;
 use crate::remote::*;
-use crate::serialize::*;
+use crate::serializer::*;
 use crate::setup::*;
 use crate::train::*;
 use clap::{value_parser, Arg, ArgAction, ArgMatches, Command};
